pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-auth>=2.38.0",
    "gspread>=6.1.4",
    "requests>=2.32.3",
]
